refactor: 초기화 hook 분리
